+++ examples/9_fpga/tb/tb.vhd
@@ -1,9 +1,6 @@
--- libstored, distributed debuggable data stores.
--- Copyright (C) 2020-2023  Jochem Rutgers
+-- SPDX-FileCopyrightText: 2020-2023 Jochem Rutgers
 --
--- This Source Code Form is subject to the terms of the Mozilla Public
--- License, v. 2.0. If a copy of the MPL was not distributed with this
--- file, You can obtain one at https://mozilla.org/MPL/2.0/.
+-- SPDX-License-Identifier: CC0-1.0
 
 library ieee;
 use ieee.std_logic_1164.all;