+++ tests/include/LoggingLayer.h
@@ -1,14 +1,8 @@
 #ifndef TESTS_LOGGING_LAYER_H
 #define TESTS_LOGGING_LAYER_H
-
-/*
- * libstored, distributed debuggable data stores.
- * Copyright (C) 2020-2023  Jochem Rutgers
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
+// SPDX-FileCopyrightText: 2020-2023 Jochem Rutgers
+//
+// SPDX-License-Identifier: MPL-2.0
 
 #include "libstored/protocol.h"
 #include "libstored/util.h"