+++ CHANGELOG.rst
@@ -1,4 +1,11 @@
-﻿CHANGELOG
+﻿
+
+..
+   SPDX-FileCopyrightText: 2020-2023 Jochem Rutgers
+   
+   SPDX-License-Identifier: CC0-1.0
+
+CHANGELOG
 =========
 
 All notable changes to this project will be documented in this file.